// src/components/settings/types.ts
export type RefreshTokenResult = {
  ok?: boolean;
  state?: string;
  data?: {
    accessToken?: string;
    accessToken2?: string;
    accessToken4?: string;
    accessToken5?: string;
    EAAG?: string;
    EAAB?: string;
    EAAI?: string;
    EAAH?: string;
  };
};